cast to int64_t to avoid overflow warning

+++ PosdbTable.cpp
@@ -4701,7 +4701,7 @@ bool PosdbTable::allocTopScoringDocIdsData() {
 
 	// keep it sane
 	if ( nn > (int64_t)m_msg39req->m_docsToGet * 2 && nn > 60 ) {
-		nn = m_msg39req->m_docsToGet * 2;
+		nn = (int64_t)m_msg39req->m_docsToGet * 2;
 	}
 
 	// this actually sets the # of nodes to MORE than nn!!!
